Docker&Streamlit/opt/requirements.txt
streamlit==1.17.0
pandas==1.4.1
torch==2.0.1
torchvision==0.15.2
python-doctr[torch]==0.6.0
altair==4.2.0
rapidfuzz==2.15.1